+++ extensions/fastly/src/tls.tsx
@@ -1,0 +1,4 @@
+import { TlsList } from "./views/tls-list";
+export default function Command() {
+  return <TlsList />;
+}